Adding basic compound interest calculator js files
I initially started this project as a way of learning the ins-and-outs of classic JS in order to build a Chrome extension for one of my classes. All my previous experience has been with frameworks in Node.js. So this is an opportunity to learn pure JS.

I thought what I built was pretty cool, so I elected to put it into a repository to continue documenting the build progress.

Currently the app is very simple, but I intend to add features to account for variability in one's salary and market conditions, for example. More to come!